langchain_scraper/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
# 
azure-functions
azurefunctions-extensions-http-fastapi #as suggested by the microsoft documentation to enable HTTP streams (https://learn.microsoft.com/en-us/azure/azure-functions/functions-reference-python?pivots=python-mode-decorators&tabs=get-started%2Casgi%2Capplication-level#triggers-and-inputs) t
langchain
beautifulsoup4
requests
azure-storage-blob
# playwright
# tqdm